requirements.txt
python-telegram-bot~=21.6
requests~=2.31
PyYAML~=6.0
publicsuffix2
